backend/requirements.txt
flask
python-dotenv
opencv-python
numpy
inference-sdk
openai-whisper
pyaudio
websockets
